Fix to fall_down
Addresses issue []. 

Due to minetest.line_of_sight returning the coordinates of the first solid node it intercepts, the bot will end its fall inside the solid node. By raising the y value of pos3 by 1, this cause the bot to land on top of the solid node. 

It also extends the maximum range the bot will drop by one so it matches the 10 node drop in the documentation. (x-1)-(x-10) = 9, while (x-1)-(x-11) = 10.